Updated output files with the correct output
LAL contained a bug (that has been fixed a few minutes ago). The bug
was that the root of a rooted tree was not unset after being removed.
This means that the method 'has_root()' returned true even after the
root had been removed. This also affected how the trees are printed to
standard output.